:label: [rename] assets의 sound 이름 변경 + ui>screen에 패키지 추가
